Add XDP loader library and CLI integration
