+++ udcn2/Cargo.toml
@@ -5,6 +5,9 @@ edition = "2021"
 
 license.workspace = true
 
+[lib]
+path = "src/lib.rs"
+
 [dependencies]
 udcn2-common = { path = "../udcn2-common", features = ["user"] }
